fix integer overflow in DropboxManagerService
Integer overflows if a device has a large storage than 1TB.
(mStatFs.getBlockCount() * reservePercent > Integer.MAX_VALUE)

Test: CtsIncidentHostTestCases
com.android.server.cts.ErrorsTest#testNativeCrash
com.android.server.cts.ErrorsTest#testThrowException

+++ services/core/java/com/android/server/DropBoxManagerService.java
@@ -862,9 +862,9 @@ public final class DropBoxManagerService extends SystemService {
                 throw new IOException("Can't restat: " + mDropBoxDir);
             }
             int available = mStatFs.getAvailableBlocks();
-            int nonreserved = available - mStatFs.getBlockCount() * reservePercent / 100;
+            int nonreserved = available - (int)((long)mStatFs.getBlockCount() * reservePercent / 100);
             int maximum = quotaKb * 1024 / mBlockSize;
-            mCachedQuotaBlocks = Math.min(maximum, Math.max(0, nonreserved * quotaPercent / 100));
+            mCachedQuotaBlocks = Math.min(maximum, Math.max(0, (int)((long)nonreserved * quotaPercent / 100)));
             mCachedQuotaUptimeMillis = uptimeMillis;
         }
